Network.AWS.Rekognition.DeleteCollection

Description

Deletes the specified collection. Note that this operation removes all faces in the collection. For an example, see example1 .

This operation requires permissions to perform the rekognition:DeleteCollection action.

Creating a Request

deleteCollection

Arguments

:: Text
dcCollectionId

-> DeleteCollection

Creates a value of DeleteCollection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcCollectionId - ID of the collection to delete.

data DeleteCollection

See: deleteCollection smart constructor.

Request Lenses

dcCollectionId :: Lens' DeleteCollection Text

ID of the collection to delete.

Destructuring the Response

deleteCollectionResponse

Arguments

:: Int
dcrsResponseStatus

-> DeleteCollectionResponse

Creates a value of DeleteCollectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcrsStatusCode - HTTP status code that indicates the result of the operation.

dcrsResponseStatus - -- | The response status code.

data DeleteCollectionResponse

See: deleteCollectionResponse smart constructor.

Response Lenses

dcrsStatusCode :: Lens' DeleteCollectionResponse (Maybe Natural)

HTTP status code that indicates the result of the operation.

dcrsResponseStatus :: Lens' DeleteCollectionResponse Int

- | The response status code.