Network.AWS.Rekognition.DeleteFaces

Description

Deletes faces from a collection. You specify a collection ID and an array of face IDs to remove from the collection.

This operation requires permissions to perform the rekognition:DeleteFaces action.

Creating a Request

deleteFaces

Arguments

:: Text
dfCollectionId

-> NonEmpty Text
dfFaceIds

-> DeleteFaces

Creates a value of DeleteFaces with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfCollectionId - Collection from which to remove the specific faces.

dfFaceIds - An array of face IDs to delete.

data DeleteFaces

See: deleteFaces smart constructor.

Request Lenses

dfCollectionId :: Lens' DeleteFaces Text

Collection from which to remove the specific faces.

dfFaceIds :: Lens' DeleteFaces (NonEmpty Text)

An array of face IDs to delete.

Destructuring the Response

deleteFacesResponse

Arguments

:: Int
drsResponseStatus

-> DeleteFacesResponse

Creates a value of DeleteFacesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsDeletedFaces - An array of strings (face IDs) of the faces that were deleted.

drsResponseStatus - -- | The response status code.

data DeleteFacesResponse

See: deleteFacesResponse smart constructor.

Response Lenses

drsDeletedFaces :: Lens' DeleteFacesResponse (Maybe (NonEmpty Text))

An array of strings (face IDs) of the faces that were deleted.

drsResponseStatus :: Lens' DeleteFacesResponse Int

- | The response status code.