Network.AWS.Rekognition.IndexFaces

Description

Detects faces in the input image and adds them to the specified collection.

Amazon Rekognition does not save the actual faces detected. Instead, the underlying detection algorithm first detects the faces in the input image, and for each face extracts facial features into a feature vector, and stores it in the back-end database. Amazon Rekognition uses feature vectors when performing face match and search operations using the and operations.

If you provide the optional externalImageID for the input image you provided, Amazon Rekognition associates this ID with all faces that it detects. When you call the operation, the response returns the external ID. You can use this external image ID to create a client-side index to associate the faces with each image. You can then use the index to find all faces in an image.

In response, the operation returns an array of metadata for all detected faces. This includes, the bounding box of the detected face, confidence value (indicating the bounding box contains a face), a face ID assigned by the service for each face that is detected and stored, and an image ID assigned by the service for the input image. If you request all facial attributes (using the detectionAttributes parameter, Amazon Rekognition returns detailed facial attributes such as facial landmarks (for example, location of eye and mount) and other facial attributes such gender. If you provide the same image, specify the same collection, and use the same external ID in the IndexFaces operation, Amazon Rekognition doesn't save duplicate face metadata.

For an example, see example2 .

This operation requires permissions to perform the rekognition:IndexFaces action.

Creating a Request

indexFaces

Arguments

:: Text
ifCollectionId

-> Image
ifImage

-> IndexFaces

Creates a value of IndexFaces with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ifExternalImageId - ID you want to assign to all the faces detected in the image.

ifDetectionAttributes - An array of facial attributes that you want to be returned. This can be the default list of attributes or all attributes. If you don't specify a value for Attributes or if you specify [DEFAULT] , the API returns the following subset of facial attributes: BoundingBox , Confidence , Pose , Quality and Landmarks . If you provide [ALL] , all facial attributes are returned but the operation will take longer to complete. If you provide both, [ALL, DEFAULT] , the service uses a logical AND operator to determine which attributes to return (in this case, all attributes).

ifCollectionId - The ID of an existing collection to which you want to add the faces that are detected in the input images.

ifImage - The input image as bytes or an S3 object.

data IndexFaces

See: indexFaces smart constructor.

Request Lenses

ifExternalImageId :: Lens' IndexFaces (Maybe Text)

ID you want to assign to all the faces detected in the image.

ifDetectionAttributes :: Lens' IndexFaces [Attribute]

An array of facial attributes that you want to be returned. This can be the default list of attributes or all attributes. If you don't specify a value for Attributes or if you specify [DEFAULT] , the API returns the following subset of facial attributes: BoundingBox , Confidence , Pose , Quality and Landmarks . If you provide [ALL] , all facial attributes are returned but the operation will take longer to complete. If you provide both, [ALL, DEFAULT] , the service uses a logical AND operator to determine which attributes to return (in this case, all attributes).

ifCollectionId :: Lens' IndexFaces Text

The ID of an existing collection to which you want to add the faces that are detected in the input images.

ifImage :: Lens' IndexFaces Image

The input image as bytes or an S3 object.

Destructuring the Response

indexFacesResponse

Arguments

:: Int
ifrsResponseStatus

-> IndexFacesResponse

Creates a value of IndexFacesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ifrsFaceRecords - An array of faces detected and added to the collection. For more information, see 'howitworks-index-faces' .

ifrsOrientationCorrection - The orientation of the input image (counterclockwise direction). If your application displays the image, you can use this value to correct image orientation. The bounding box coordinates returned in FaceRecords represent face locations before the image orientation is corrected.

ifrsResponseStatus - -- | The response status code.

data IndexFacesResponse

See: indexFacesResponse smart constructor.

Response Lenses

ifrsFaceRecords :: Lens' IndexFacesResponse [FaceRecord]

An array of faces detected and added to the collection. For more information, see 'howitworks-index-faces' .

ifrsOrientationCorrection :: Lens' IndexFacesResponse (Maybe OrientationCorrection)

The orientation of the input image (counterclockwise direction). If your application displays the image, you can use this value to correct image orientation. The bounding box coordinates returned in FaceRecords represent face locations before the image orientation is corrected.

ifrsResponseStatus :: Lens' IndexFacesResponse Int

- | The response status code.