Network.AWS.Rekognition.ListCollections

Description

Returns list of collection IDs in your account. If the result is truncated, the response also provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

For an example, see example1 .

This operation requires permissions to perform the rekognition:ListCollections action.

This operation returns paginated results.

Creating a Request

listCollections :: ListCollections

Creates a value of ListCollections with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcNextToken - Pagination token from the previous response.

lcMaxResults - Maximum number of collection IDs to return.

data ListCollections

See: listCollections smart constructor.

Request Lenses

lcNextToken :: Lens' ListCollections (Maybe Text)

Pagination token from the previous response.

lcMaxResults :: Lens' ListCollections (Maybe Natural)

Maximum number of collection IDs to return.

Destructuring the Response

listCollectionsResponse

Arguments

:: Int
lcrsResponseStatus

-> ListCollectionsResponse

Creates a value of ListCollectionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lcrsCollectionIds - An array of collection IDs.

lcrsNextToken - If the result is truncated, the response provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

lcrsResponseStatus - -- | The response status code.

data ListCollectionsResponse

See: listCollectionsResponse smart constructor.

Response Lenses

lcrsCollectionIds :: Lens' ListCollectionsResponse [Text]

An array of collection IDs.

lcrsNextToken :: Lens' ListCollectionsResponse (Maybe Text)

If the result is truncated, the response provides a NextToken that you can use in the subsequent request to fetch the next set of collection IDs.

lcrsResponseStatus :: Lens' ListCollectionsResponse Int

- | The response status code.