Network.AWS.Rekognition.ListFaces

Description

Returns metadata for faces in the specified collection. This metadata includes information such as the bounding box coordinates, the confidence (that the bounding box contains a face), and face ID. For an example, see example3 .

This operation requires permissions to perform the rekognition:ListFaces action.

This operation returns paginated results.

Creating a Request

listFaces

Arguments

:: Text
lfCollectionId

-> ListFaces

Creates a value of ListFaces with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfNextToken - If the previous response was incomplete (because there is more data to retrieve), Amazon Rekognition returns a pagination token in the response. You can use this pagination token to retrieve the next set of faces.

lfMaxResults - Maximum number of faces to return.

lfCollectionId - ID of the collection from which to list the faces.

data ListFaces

See: listFaces smart constructor.

Request Lenses

lfNextToken :: Lens' ListFaces (Maybe Text)

If the previous response was incomplete (because there is more data to retrieve), Amazon Rekognition returns a pagination token in the response. You can use this pagination token to retrieve the next set of faces.

lfMaxResults :: Lens' ListFaces (Maybe Natural)

Maximum number of faces to return.

lfCollectionId :: Lens' ListFaces Text

ID of the collection from which to list the faces.

Destructuring the Response

listFacesResponse

Arguments

:: Int
lfrsResponseStatus

-> ListFacesResponse

Creates a value of ListFacesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfrsNextToken - If the response is truncated, Amazon Rekognition returns this token that you can use in the subsequent request to retrieve the next set of faces.

lfrsFaces - An array of Face objects.

lfrsResponseStatus - -- | The response status code.

data ListFacesResponse

See: listFacesResponse smart constructor.

Response Lenses

lfrsNextToken :: Lens' ListFacesResponse (Maybe Text)

If the response is truncated, Amazon Rekognition returns this token that you can use in the subsequent request to retrieve the next set of faces.

lfrsFaces :: Lens' ListFacesResponse [Face]

An array of Face objects.

lfrsResponseStatus :: Lens' ListFacesResponse Int

- | The response status code.