Network.AWS.Rekognition.SearchFacesByImage

Description

For a given input image, first detects the largest face in the image, and then searches the specified collection for matching faces. The operation compares the features of the input face with faces in the specified collection.

The response returns an array of faces that match, ordered by similarity score with the highest similarity first. More specifically, it is an array of metadata for each face match found. Along with the metadata, the response also includes a similarity indicating how similar the face is to the input face. In the response, the operation also returns the bounding box (and a confidence level that the bounding box contains a face) of the face that Amazon Rekognition used for the input image.

For an example, see example3 .

This operation requires permissions to perform the rekognition:SearchFacesByImage action.

Creating a Request

searchFacesByImage

Arguments

:: Text
sfbiCollectionId

-> Image
sfbiImage

-> SearchFacesByImage

Creates a value of SearchFacesByImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sfbiFaceMatchThreshold - (Optional) Specifies the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%.

sfbiMaxFaces - Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

sfbiCollectionId - ID of the collection to search.

sfbiImage - The input image as bytes or an S3 object.

data SearchFacesByImage

See: searchFacesByImage smart constructor.

Request Lenses

sfbiFaceMatchThreshold :: Lens' SearchFacesByImage (Maybe Double)

(Optional) Specifies the minimum confidence in the face match to return. For example, don't return any matches where confidence in matches is less than 70%.

sfbiMaxFaces :: Lens' SearchFacesByImage (Maybe Natural)

Maximum number of faces to return. The operation returns the maximum number of faces with the highest confidence in the match.

sfbiCollectionId :: Lens' SearchFacesByImage Text

ID of the collection to search.

sfbiImage :: Lens' SearchFacesByImage Image

The input image as bytes or an S3 object.

Destructuring the Response

searchFacesByImageResponse

Arguments

:: Int
sfbirsResponseStatus

-> SearchFacesByImageResponse

Creates a value of SearchFacesByImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sfbirsFaceMatches - An array of faces that match the input face, along with the confidence in the match.

sfbirsSearchedFaceBoundingBox - The bounding box around the face in the input image that Amazon Rekognition used for the search.

sfbirsSearchedFaceConfidence - The level of confidence that the searchedFaceBoundingBox , contains a face.

sfbirsResponseStatus - -- | The response status code.

data SearchFacesByImageResponse

See: searchFacesByImageResponse smart constructor.

Response Lenses

sfbirsFaceMatches :: Lens' SearchFacesByImageResponse [FaceMatch]

An array of faces that match the input face, along with the confidence in the match.

sfbirsSearchedFaceBoundingBox :: Lens' SearchFacesByImageResponse (Maybe BoundingBox)

The bounding box around the face in the input image that Amazon Rekognition used for the search.

sfbirsSearchedFaceConfidence :: Lens' SearchFacesByImageResponse (Maybe Double)

The level of confidence that the searchedFaceBoundingBox , contains a face.

sfbirsResponseStatus :: Lens' SearchFacesByImageResponse Int

- | The response status code.