Network.AWS.Rekognition.Types

Description

Service Configuration

rekognition :: Service

API version 2016-06-27 of the Amazon Rekognition SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to perform the action.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

Input parameter violated a constraint. Validate your parameter before calling the API operation again.

_InvalidImageFormatException :: AsError a => Getting (First ServiceError) a ServiceError

The provided image format is not supported.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A collection with the specified ID already exists.

_InvalidS3ObjectException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Rekognition is unable to access the S3 object specified in the request.

_ProvisionedThroughputExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of requests exceeded your throughput limit. If you want to increase this limit, contact Amazon Rekognition.

_ImageTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The input image size exceeds the allowed limit. For more information, see limits .

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Rekognition is temporarily unable to process the request. Try your call again.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Rekognition experienced a service issue. Try your call again.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Collection specified in the request is not found.

_InvalidPaginationTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Pagination token in the request is not valid.

Attribute

data Attribute

Constructors

All

Default

EmotionName

data EmotionName

Constructors

Angry

Calm

Confused

Disgusted

Happy

Sad

Surprised

Unknown

GenderType

data GenderType

Constructors

Female

Male

LandmarkType

data LandmarkType

Constructors

EyeLeft

EyeRight

LeftEyeDown

LeftEyeLeft

LeftEyeRight

LeftEyeUp

LeftEyebrowLeft

LeftEyebrowRight

LeftEyebrowUp

LeftPupil

MouthDown

MouthLeft

MouthRight

MouthUp

Nose

NoseLeft

NoseRight

RightEyeDown

RightEyeLeft

RightEyeRight

RightEyeUp

RightEyebrowLeft

RightEyebrowRight

RightEyebrowUp

RightPupil

OrientationCorrection

data OrientationCorrection

Constructors

Rotate0

Rotate180

Rotate270

Rotate90

AgeRange

data AgeRange

Structure containing the estimated age range, in years, for a face.

Rekognition estimates an age-range for faces detected in the input image. Estimated age ranges can overlap; a face of a 5 year old may have an estimated range of 4-6 whilst the face of a 6 year old may have an estimated range of 4-8.

See: ageRange smart constructor.

ageRange :: AgeRange

Creates a value of AgeRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arLow - The lowest estimated age.

arHigh - The highest estimated age.

arLow :: Lens' AgeRange (Maybe Natural)

The lowest estimated age.

arHigh :: Lens' AgeRange (Maybe Natural)

The highest estimated age.

Beard

data Beard

Indicates whether or not the face has a beard, and the confidence level in the determination.

See: beard smart constructor.

beard :: Beard

Creates a value of Beard with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bValue - Boolean value that indicates whether the face has beard or not.

bConfidence - Level of confidence in the determination.

bValue :: Lens' Beard (Maybe Bool)

Boolean value that indicates whether the face has beard or not.

bConfidence :: Lens' Beard (Maybe Double)

Level of confidence in the determination.

BoundingBox

data BoundingBox

Identifies the bounding box around the object or face. The left (x-coordinate) and top (y-coordinate) are coordinates representing the top and left sides of the bounding box. Note that the upper-left corner of the image is the origin (0,0).

The top and left values returned are ratios of the overall image size. For example, if the input image is 700x200 pixels, and the top-left coordinate of the bounding box is 350x50 pixels, the API returns a left value of 0.5 (350700) and a top value of 0.25 (50200).

The width and height values represent the dimensions of the bounding box as a ratio of the overall image dimension. For example, if the input image is 700x200 pixels, and the bounding box width is 70 pixels, the width returned is 0.1.

See: boundingBox smart constructor.

boundingBox :: BoundingBox

Creates a value of BoundingBox with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bbHeight - Height of the bounding box as a ratio of the overall image height.

bbLeft - Left coordinate of the bounding box as a ratio of overall image width.

bbWidth - Width of the bounding box as a ratio of the overall image width.

bbTop - Top coordinate of the bounding box as a ratio of overall image height.

bbHeight :: Lens' BoundingBox (Maybe Double)

Height of the bounding box as a ratio of the overall image height.

bbLeft :: Lens' BoundingBox (Maybe Double)

Left coordinate of the bounding box as a ratio of overall image width.

bbWidth :: Lens' BoundingBox (Maybe Double)

Width of the bounding box as a ratio of the overall image width.

bbTop :: Lens' BoundingBox (Maybe Double)

Top coordinate of the bounding box as a ratio of overall image height.

Celebrity

data Celebrity

Provides information about a celebrity recognized by the operation.

See: celebrity smart constructor.

celebrity :: Celebrity

Creates a value of Celebrity with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cMatchConfidence - The confidence, in percentage, that Rekognition has that the recognized face is the celebrity.

cURLs - An array of URLs pointing to additional information about the celebrity. If there is no additional information about the celebrity, this list is empty.

cName - The name of the celebrity.

cId - A unique identifier for the celebrity.

cFace - Provides information about the celebrity's face, such as its location on the image.

cMatchConfidence :: Lens' Celebrity (Maybe Double)

The confidence, in percentage, that Rekognition has that the recognized face is the celebrity.

cURLs :: Lens' Celebrity [Text]

An array of URLs pointing to additional information about the celebrity. If there is no additional information about the celebrity, this list is empty.

cName :: Lens' Celebrity (Maybe Text)

The name of the celebrity.

cId :: Lens' Celebrity (Maybe Text)

A unique identifier for the celebrity.

cFace :: Lens' Celebrity (Maybe ComparedFace)

Provides information about the celebrity's face, such as its location on the image.

CompareFacesMatch

data CompareFacesMatch

Provides information about a face in a target image that matches the source image face analysed by CompareFaces . The Face property contains the bounding box of the face in the target image. The Similarity property is the confidence that the source image face matches the face in the bounding box.

See: compareFacesMatch smart constructor.

compareFacesMatch :: CompareFacesMatch

Creates a value of CompareFacesMatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfmSimilarity - Level of confidence that the faces match.

cfmFace - Provides face metadata (bounding box and confidence that the bounding box actually contains a face).

cfmSimilarity :: Lens' CompareFacesMatch (Maybe Double)

Level of confidence that the faces match.

cfmFace :: Lens' CompareFacesMatch (Maybe ComparedFace)

Provides face metadata (bounding box and confidence that the bounding box actually contains a face).

ComparedFace

data ComparedFace

Provides face metadata for target image faces that are analysed by CompareFaces and RecognizeCelebrities .

See: comparedFace smart constructor.

comparedFace :: ComparedFace

Creates a value of ComparedFace with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfBoundingBox - Bounding box of the face.

cfPose - Indicates the pose of the face as determined by its pitch, roll, and yaw.

cfConfidence - Level of confidence that what the bounding box contains is a face.

cfQuality - Identifies face image brightness and sharpness.

cfLandmarks - An array of facial landmarks.

cfBoundingBox :: Lens' ComparedFace (Maybe BoundingBox)

Bounding box of the face.

cfPose :: Lens' ComparedFace (Maybe Pose)

Indicates the pose of the face as determined by its pitch, roll, and yaw.

cfConfidence :: Lens' ComparedFace (Maybe Double)

Level of confidence that what the bounding box contains is a face.

cfQuality :: Lens' ComparedFace (Maybe ImageQuality)

Identifies face image brightness and sharpness.

cfLandmarks :: Lens' ComparedFace [Landmark]

An array of facial landmarks.

ComparedSourceImageFace

data ComparedSourceImageFace

Type that describes the face Amazon Rekognition chose to compare with the faces in the target. This contains a bounding box for the selected face and confidence level that the bounding box contains a face. Note that Amazon Rekognition selects the largest face in the source image for this comparison.

See: comparedSourceImageFace smart constructor.

comparedSourceImageFace :: ComparedSourceImageFace

Creates a value of ComparedSourceImageFace with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csifBoundingBox - Bounding box of the face.

csifConfidence - Confidence level that the selected bounding box contains a face.

csifBoundingBox :: Lens' ComparedSourceImageFace (Maybe BoundingBox)

Bounding box of the face.

csifConfidence :: Lens' ComparedSourceImageFace (Maybe Double)

Confidence level that the selected bounding box contains a face.

Emotion

data Emotion

The emotions detected on the face, and the confidence level in the determination. For example, HAPPY, SAD, and ANGRY.

See: emotion smart constructor.

emotion :: Emotion

Creates a value of Emotion with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eConfidence - Level of confidence in the determination.

eType - Type of emotion detected.

eConfidence :: Lens' Emotion (Maybe Double)

Level of confidence in the determination.

eType :: Lens' Emotion (Maybe EmotionName)

Type of emotion detected.

EyeOpen

data EyeOpen

Indicates whether or not the eyes on the face are open, and the confidence level in the determination.

See: eyeOpen smart constructor.

eyeOpen :: EyeOpen

Creates a value of EyeOpen with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eoValue - Boolean value that indicates whether the eyes on the face are open.

eoConfidence - Level of confidence in the determination.

eoValue :: Lens' EyeOpen (Maybe Bool)

Boolean value that indicates whether the eyes on the face are open.

eoConfidence :: Lens' EyeOpen (Maybe Double)

Level of confidence in the determination.

Eyeglasses

data Eyeglasses

Indicates whether or not the face is wearing eye glasses, and the confidence level in the determination.

See: eyeglasses smart constructor.

eyeglasses :: Eyeglasses

Creates a value of Eyeglasses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eyeValue - Boolean value that indicates whether the face is wearing eye glasses or not.

eyeConfidence - Level of confidence in the determination.

eyeValue :: Lens' Eyeglasses (Maybe Bool)

Boolean value that indicates whether the face is wearing eye glasses or not.

eyeConfidence :: Lens' Eyeglasses (Maybe Double)

Level of confidence in the determination.

Face

data Face

Describes the face properties such as the bounding box, face ID, image ID of the input image, and external image ID that you assigned.

See: face smart constructor.

face :: Face

Creates a value of Face with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fFaceId - Unique identifier that Amazon Rekognition assigns to the face.

fBoundingBox - Bounding box of the face.

fExternalImageId - Identifier that you assign to all the faces in the input image.

fConfidence - Confidence level that the bounding box contains a face (and not a different object such as a tree).

fImageId - Unique identifier that Amazon Rekognition assigns to the input image.

fFaceId :: Lens' Face (Maybe Text)

Unique identifier that Amazon Rekognition assigns to the face.

fBoundingBox :: Lens' Face (Maybe BoundingBox)

Bounding box of the face.

fExternalImageId :: Lens' Face (Maybe Text)

Identifier that you assign to all the faces in the input image.

fConfidence :: Lens' Face (Maybe Double)

Confidence level that the bounding box contains a face (and not a different object such as a tree).

fImageId :: Lens' Face (Maybe Text)

Unique identifier that Amazon Rekognition assigns to the input image.

FaceDetail

data FaceDetail

Structure containing attributes of the face that the algorithm detected.

See: faceDetail smart constructor.

faceDetail :: FaceDetail

Creates a value of FaceDetail with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fdAgeRange - The estimated age range, in years, for the face. Low represents the lowest estimated age and High represents the highest estimated age.

fdSunglasses - Indicates whether or not the face is wearing sunglasses, and the confidence level in the determination.

fdMouthOpen - Indicates whether or not the mouth on the face is open, and the confidence level in the determination.

fdBoundingBox - Bounding box of the face.

fdEmotions - The emotions detected on the face, and the confidence level in the determination. For example, HAPPY, SAD, and ANGRY.

fdEyesOpen - Indicates whether or not the eyes on the face are open, and the confidence level in the determination.

fdPose - Indicates the pose of the face as determined by its pitch, roll, and yaw.

fdConfidence - Confidence level that the bounding box contains a face (and not a different object such as a tree).

fdGender - Gender of the face and the confidence level in the determination.

fdQuality - Identifies image brightness and sharpness.

fdEyeglasses - Indicates whether or not the face is wearing eye glasses, and the confidence level in the determination.

fdBeard - Indicates whether or not the face has a beard, and the confidence level in the determination.

fdMustache - Indicates whether or not the face has a mustache, and the confidence level in the determination.

fdSmile - Indicates whether or not the face is smiling, and the confidence level in the determination.

fdLandmarks - Indicates the location of landmarks on the face.

fdAgeRange :: Lens' FaceDetail (Maybe AgeRange)

The estimated age range, in years, for the face. Low represents the lowest estimated age and High represents the highest estimated age.

fdSunglasses :: Lens' FaceDetail (Maybe Sunglasses)

Indicates whether or not the face is wearing sunglasses, and the confidence level in the determination.

fdMouthOpen :: Lens' FaceDetail (Maybe MouthOpen)

Indicates whether or not the mouth on the face is open, and the confidence level in the determination.

fdBoundingBox :: Lens' FaceDetail (Maybe BoundingBox)

Bounding box of the face.

fdEmotions :: Lens' FaceDetail [Emotion]

The emotions detected on the face, and the confidence level in the determination. For example, HAPPY, SAD, and ANGRY.

fdEyesOpen :: Lens' FaceDetail (Maybe EyeOpen)

Indicates whether or not the eyes on the face are open, and the confidence level in the determination.

fdPose :: Lens' FaceDetail (Maybe Pose)

Indicates the pose of the face as determined by its pitch, roll, and yaw.

fdConfidence :: Lens' FaceDetail (Maybe Double)

Confidence level that the bounding box contains a face (and not a different object such as a tree).

fdGender :: Lens' FaceDetail (Maybe Gender)

Gender of the face and the confidence level in the determination.

fdQuality :: Lens' FaceDetail (Maybe ImageQuality)

Identifies image brightness and sharpness.

fdEyeglasses :: Lens' FaceDetail (Maybe Eyeglasses)

Indicates whether or not the face is wearing eye glasses, and the confidence level in the determination.

fdBeard :: Lens' FaceDetail (Maybe Beard)

Indicates whether or not the face has a beard, and the confidence level in the determination.

fdMustache :: Lens' FaceDetail (Maybe Mustache)

Indicates whether or not the face has a mustache, and the confidence level in the determination.

fdSmile :: Lens' FaceDetail (Maybe Smile)

Indicates whether or not the face is smiling, and the confidence level in the determination.

fdLandmarks :: Lens' FaceDetail [Landmark]

Indicates the location of landmarks on the face.

FaceMatch

data FaceMatch

Provides face metadata. In addition, it also provides the confidence in the match of this face with the input face.

See: faceMatch smart constructor.

faceMatch :: FaceMatch

Creates a value of FaceMatch with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fmSimilarity - Confidence in the match of this face with the input face.

fmFace - Describes the face properties such as the bounding box, face ID, image ID of the source image, and external image ID that you assigned.

fmSimilarity :: Lens' FaceMatch (Maybe Double)

Confidence in the match of this face with the input face.

fmFace :: Lens' FaceMatch (Maybe Face)

Describes the face properties such as the bounding box, face ID, image ID of the source image, and external image ID that you assigned.

FaceRecord

data FaceRecord

Object containing both the face metadata (stored in the back-end database) and facial attributes that are detected but aren't stored in the database.

See: faceRecord smart constructor.

faceRecord :: FaceRecord

Creates a value of FaceRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

frFaceDetail - Structure containing attributes of the face that the algorithm detected.

frFace - Describes the face properties such as the bounding box, face ID, image ID of the input image, and external image ID that you assigned.

frFaceDetail :: Lens' FaceRecord (Maybe FaceDetail)

Structure containing attributes of the face that the algorithm detected.

frFace :: Lens' FaceRecord (Maybe Face)

Describes the face properties such as the bounding box, face ID, image ID of the input image, and external image ID that you assigned.

Gender

data Gender

Gender of the face and the confidence level in the determination.

See: gender smart constructor.

gender :: Gender

Creates a value of Gender with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gValue - Gender of the face.

gConfidence - Level of confidence in the determination.

gValue :: Lens' Gender (Maybe GenderType)

Gender of the face.

gConfidence :: Lens' Gender (Maybe Double)

Level of confidence in the determination.

Image

data Image

Provides the input image either as bytes or an S3 object.

You pass image bytes to a Rekognition API operation by using the Bytes property. For example, you would use the Bytes property to pass an image loaded from a local file system. Image bytes passed by using the Bytes property must be base64-encoded. Your code may not need to encode image bytes if you are using an AWS SDK to call Rekognition API operations. For more information, see example4 .

You pass images stored in an S3 bucket to a Rekognition API operation by using the S3Object property. Images stored in an S3 bucket do not need to be base64-encoded.

The region for the S3 bucket containing the S3 object must match the region you use for Amazon Rekognition operations.

If you use the Amazon CLI to call Amazon Rekognition operations, passing image bytes using the Bytes property is not supported. You must first upload the image to an Amazon S3 bucket and then call the operation using the S3Object property.

For Amazon Rekognition to process an S3 object, the user must have permission to access the S3 object. For more information, see 'manage-access-resource-policies' .

See: image smart constructor.

image :: Image

Creates a value of Image with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iS3Object - Identifies an S3 object as the image source.

iBytes - Blob of image bytes up to 5 MBs.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

iS3Object :: Lens' Image (Maybe S3Object)

Identifies an S3 object as the image source.

iBytes :: Lens' Image (Maybe ByteString)

Blob of image bytes up to 5 MBs.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

ImageQuality

data ImageQuality

Identifies face image brightness and sharpness.

See: imageQuality smart constructor.

imageQuality :: ImageQuality

Creates a value of ImageQuality with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

iqSharpness - Value representing sharpness of the face. The service returns a value between 0 and 100 (inclusive). A higher value indicates a sharper face image.

iqBrightness - Value representing brightness of the face. The service returns a value between 0 and 100 (inclusive). A higher value indicates a brighter face image.

iqSharpness :: Lens' ImageQuality (Maybe Double)

Value representing sharpness of the face. The service returns a value between 0 and 100 (inclusive). A higher value indicates a sharper face image.

iqBrightness :: Lens' ImageQuality (Maybe Double)

Value representing brightness of the face. The service returns a value between 0 and 100 (inclusive). A higher value indicates a brighter face image.

Label

data Label

Structure containing details about the detected label, including name, and level of confidence.

See: label smart constructor.

label :: Label

Creates a value of Label with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lConfidence - Level of confidence.

lName - The name (label) of the object.

lConfidence :: Lens' Label (Maybe Double)

Level of confidence.

lName :: Lens' Label (Maybe Text)

The name (label) of the object.

Landmark

data Landmark

Indicates the location of the landmark on the face.

See: landmark smart constructor.

landmark :: Landmark

Creates a value of Landmark with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lType - Type of the landmark.

lX - x-coordinate from the top left of the landmark expressed as the ration of the width of the image. For example, if the images is 700x200 and the x-coordinate of the landmark is at 350 pixels, this value is 0.5.

lY - y-coordinate from the top left of the landmark expressed as the ration of the height of the image. For example, if the images is 700x200 and the y-coordinate of the landmark is at 100 pixels, this value is 0.5.

lType :: Lens' Landmark (Maybe LandmarkType)

Type of the landmark.

lX :: Lens' Landmark (Maybe Double)

x-coordinate from the top left of the landmark expressed as the ration of the width of the image. For example, if the images is 700x200 and the x-coordinate of the landmark is at 350 pixels, this value is 0.5.

lY :: Lens' Landmark (Maybe Double)

y-coordinate from the top left of the landmark expressed as the ration of the height of the image. For example, if the images is 700x200 and the y-coordinate of the landmark is at 100 pixels, this value is 0.5.

ModerationLabel

data ModerationLabel

Provides information about a single type of moderated content found in an image. Each type of moderated content has a label within a hierarchical taxonomy. For more information, see 'image-moderation' .

See: moderationLabel smart constructor.

moderationLabel :: ModerationLabel

Creates a value of ModerationLabel with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mlConfidence - Specifies the confidence that Amazon Rekognition has that the label has been correctly identified. If you don't specify the MinConfidence parameter in the call to DetectModerationLabels , the operation returns labels with a confidence value greater than or equal to 50 percent.

mlName - The label name for the type of content detected in the image.

mlParentName - The name for the parent label. Labels at the top-level of the hierarchy have the parent label "" .

mlConfidence :: Lens' ModerationLabel (Maybe Double)

Specifies the confidence that Amazon Rekognition has that the label has been correctly identified. If you don't specify the MinConfidence parameter in the call to DetectModerationLabels , the operation returns labels with a confidence value greater than or equal to 50 percent.

mlName :: Lens' ModerationLabel (Maybe Text)

The label name for the type of content detected in the image.

mlParentName :: Lens' ModerationLabel (Maybe Text)

The name for the parent label. Labels at the top-level of the hierarchy have the parent label "" .

MouthOpen

data MouthOpen

Indicates whether or not the mouth on the face is open, and the confidence level in the determination.

See: mouthOpen smart constructor.

mouthOpen :: MouthOpen

Creates a value of MouthOpen with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

moValue - Boolean value that indicates whether the mouth on the face is open or not.

moConfidence - Level of confidence in the determination.

moValue :: Lens' MouthOpen (Maybe Bool)

Boolean value that indicates whether the mouth on the face is open or not.

moConfidence :: Lens' MouthOpen (Maybe Double)

Level of confidence in the determination.

Mustache

data Mustache

Indicates whether or not the face has a mustache, and the confidence level in the determination.

See: mustache smart constructor.

mustache :: Mustache

Creates a value of Mustache with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mValue - Boolean value that indicates whether the face has mustache or not.

mConfidence - Level of confidence in the determination.

mValue :: Lens' Mustache (Maybe Bool)

Boolean value that indicates whether the face has mustache or not.

mConfidence :: Lens' Mustache (Maybe Double)

Level of confidence in the determination.

Pose

data Pose

Indicates the pose of the face as determined by its pitch, roll, and yaw.

See: pose smart constructor.

pose :: Pose

Creates a value of Pose with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pYaw - Value representing the face rotation on the yaw axis.

pRoll - Value representing the face rotation on the roll axis.

pPitch - Value representing the face rotation on the pitch axis.

pYaw :: Lens' Pose (Maybe Double)

Value representing the face rotation on the yaw axis.

pRoll :: Lens' Pose (Maybe Double)

Value representing the face rotation on the roll axis.

pPitch :: Lens' Pose (Maybe Double)

Value representing the face rotation on the pitch axis.

S3Object

data S3Object

Provides the S3 bucket name and object name.

The region for the S3 bucket containing the S3 object must match the region you use for Amazon Rekognition operations.

For Amazon Rekognition to process an S3 object, the user must have permission to access the S3 object. For more information, see 'manage-access-resource-policies' .

See: s3Object smart constructor.

s3Object :: S3Object

Creates a value of S3Object with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

soBucket - Name of the S3 bucket.

soName - S3 object key name.

soVersion - If the bucket is versioning enabled, you can specify the object version.

soBucket :: Lens' S3Object (Maybe Text)

Name of the S3 bucket.

soName :: Lens' S3Object (Maybe Text)

S3 object key name.

soVersion :: Lens' S3Object (Maybe Text)

If the bucket is versioning enabled, you can specify the object version.

Smile

data Smile

Indicates whether or not the face is smiling, and the confidence level in the determination.

See: smile smart constructor.

smile :: Smile

Creates a value of Smile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

smiValue - Boolean value that indicates whether the face is smiling or not.

smiConfidence - Level of confidence in the determination.

smiValue :: Lens' Smile (Maybe Bool)

Boolean value that indicates whether the face is smiling or not.

smiConfidence :: Lens' Smile (Maybe Double)

Level of confidence in the determination.

Sunglasses

data Sunglasses

Indicates whether or not the face is wearing sunglasses, and the confidence level in the determination.

See: sunglasses smart constructor.

sunglasses :: Sunglasses

Creates a value of Sunglasses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sValue - Boolean value that indicates whether the face is wearing sunglasses or not.

sConfidence - Level of confidence in the determination.

sValue :: Lens' Sunglasses (Maybe Bool)

Boolean value that indicates whether the face is wearing sunglasses or not.

sConfidence :: Lens' Sunglasses (Maybe Double)

Level of confidence in the determination.